+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Changed
+
+- super+backspace is now an alias for ctrl+u in Input and TextArea
+
 ## [8.2.7] - 2026-05-19 
 
 ### Added